requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
django==4.2.26
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.11
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
stripe==13.2.0
typing-extensions==4.13.2
urllib3==2.2.3
